requirements.txt
alembic==1.11.1
anyio==3.7.0
asgiref==3.7.2
asyncpg==0.27.0
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.7
cryptography==41.0.1
Django==4.2.1
djangorestframework==3.14.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.96.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.3
mirakuru==2.5.1
packaging==23.1
passlib==1.7.4
platformdirs==3.5.1
pluggy==1.0.0
port-for==0.6.3
psutil==5.9.5
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-pool==3.1.7
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
pypng==0.20220715.0
pytest==7.3.1
pytest-postgresql==5.0.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqladmin==0.11.0
SQLAlchemy==2.0.15
sqlalchemy-json==0.5.0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.2
uvicorn==0.22.0
whitenoise==6.4.0
WTForms==3.0.1
zeep==4.2.1
